youtube-mcp-server/pyproject.toml
[project]
name = "youtube-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.1",
    "python-dotenv>=1.1.0",
    "youtube-transcript-api>=1.0.3",
]
